Improve tests and example
