+++ cookbook/mixed-transitions.html
@@ -12,11 +12,15 @@
         <script src="https://unpkg.com/@ionic/core@4.0.0-beta.2/dist/ionic.js"></script>
         <link rel="stylesheet" href="https://unpkg.com/@ionic/core@4.0.0-beta.2/css/ionic.min.css"/>
         <style>
-            .fade-enter-active, .fade-leave-active {
-                transition: all 1s ease;
+            /* hide the nasty black BG on ios */
+            .show-decor .nav-decor {
+                background: transparent !important;
             }
-            .fade-enter, .fade-leave-to {
-                opacity: 0
+            .slide-enter-active, .slide-leave-active {
+                transition: margin 1s ease;
+            }
+            .slide-enter, .slide-leave-to {
+                margin-top: 1000px;
             }
         </style>
     </head>
@@ -41,16 +45,16 @@
             })
 
             const Home = {
-                template: `<ion-page class="ion-page">
+                template: `<transition name="slide"><ion-page class="ion-page">
                     <toolbar title="Home"/>
                     <ion-content class="ion-content" padding>
                         <router-link to="/page">Go to page</router-link>
                     </ion-content>
-                </ion-page>`
+                </ion-page></transition>`
             }
 
             const Page = {
-                template: `<transition name="fade">
+                template: `<transition name="slide">
                     <ion-page class="ion-page">
                         <toolbar title="Page" backURL="/"/>
                         <ion-content class="ion-content" padding>
