+++ test/ion-vue-router.spec.js
@@ -1,7 +1,6 @@
 import Vue from 'vue'
 import Router from '../src/router.js'
 import IonVueRouter from '../src/components/ion-vue-router.vue'
-import IonVueRouterTransitionless from '../src/components/ion-vue-router-transitionless.vue'
 
 describe('IonVueRouter', () => {
   Vue.use(Router)
@@ -56,29 +55,6 @@ describe('IonVueRouter', () => {
     expect(component.bindCss).toBeFalsy()
     expect(component.animated).toBeTruthy()
     expect(component.name).toBe('default')
-  })
-
-  it('Catches back button click event', () => {
-    const constructor = Vue.extend(IonVueRouterTransitionless)
-    const component = new constructor({ router: new Router({ mode: 'abstract' }) })
-
-    expect(component.catchIonicGoBack({})).toBeFalsy()
-
-    component.$router.push('/')
-    component.$router.push('/foo')
-    expect(component.$route.fullPath).toBe('/foo')
-
-    // Go back
-    component.catchIonicGoBack(mockBackEvent('/'))
-    expect(component.$route.fullPath).toBe('/')
-
-    // Should not go back
-    component.catchIonicGoBack(mockBackEvent())
-    expect(component.$route.fullPath).toBe('/')
-
-    // Go back to default route
-    component.catchIonicGoBack(mockBackEvent('/bar'))
-    expect(component.$route.fullPath).toBe('/bar')
   })
 
   it('Transitions correctly', () => {
@@ -190,18 +166,5 @@ function mockIonRouterOutlet() {
         })
       })
     },
-  }
-}
-
-function mockBackEvent(route) {
-  return {
-    target: {
-      closest() {
-        return {
-          defaultHref: route,
-        }
-      },
-    },
-    preventDefault() {},
   }
 }
